Reorganize code in folder
