+++ android/brave_java_sources.gni
@@ -68,9 +68,9 @@ brave_java_sources = [
   "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/NTPBackgroundImagesBridge.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/RewardsBottomSheetDialogFragment.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/SuperReferralShareDialogFragment.java",
-  "../../brave/android/java/org/chromium/chrome/browser/crypto_widget/CryptoWidgetBottomSheetDialogFragment.java",
-  "../../brave/android/java/org/chromium/chrome/browser/crypto_widget/BlankFragment.java",
-  "../../brave/android/java/org/chromium/chrome/browser/crypto_widget/CryptoWidgetTabAdapter.java",
+  "../../brave/android/java/org/chromium/chrome/browser/widget/crypto/binance/CryptoWidgetBottomSheetDialogFragment.java",
+  "../../brave/android/java/org/chromium/chrome/browser/widget/crypto/binance/BlankFragment.java",
+  "../../brave/android/java/org/chromium/chrome/browser/widget/crypto/binance/CryptoWidgetTabAdapter.java",
   "../../brave/android/java/org/chromium/chrome/browser/omnibox/suggestions/editurl/BraveEditUrlSuggestionProcessor.java",
   "../../brave/android/java/org/chromium/chrome/browser/onboarding/BraveRewardsOnboardingFragment.java",
   "../../brave/android/java/org/chromium/chrome/browser/onboarding/BraveRewardsService.java",
