+++ android/java/org/chromium/chrome/browser/widget/crypto/binance/CryptoWidgetBottomSheetDialogFragment.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.crypto_widget;
+package org.chromium.chrome.browser.widget.crypto.binance;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
